overclock가격 조정, overclockpanel 껐다켜면 0되는 문제 수정 전

+++ Assets/Script/BtcPanel.cs
@@ -164,7 +164,7 @@ public class BtcPanel : MonoBehaviour
     {
         try
         {
-            float btcForBuy = scriptGameSystem.currentBtc;
+            double btcForBuy = scriptGameSystem.currentBtc;
             UInt64 moneyToGet = Convert.ToUInt64(scriptGameSystem.currentBtcPrice* btcForBuy);
 
             if (btcForBuy < 0.00000001)
